% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_period.R
\name{adjust_period}
\alias{adjust_period}
\title{Adjust Period to Monthly Equivalent}
\usage{
adjust_period(d)
}
\arguments{
\item{d}{A data frame containing the following logical indicators: \code{month}, \code{quarter}, \code{annual}.}
}
\value{
An integer vector representing the adjusted period in months.
}
\description{
This function adjusts a given period value to its equivalent in months based on the provided frequency indicators from the coding.
}
\examples{
d$period.month <- adjust_period(d)

}
