% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_shock_periodicity.R
\name{adjust_shock_periodicity}
\alias{adjust_shock_periodicity}
\title{Adjust shock_size based on the periodicity of interest rate.}
\usage{
adjust_shock_periodicity(d, shock_size)
}
\arguments{
\item{d}{An observation (i.e., a row) from a dataframe obtained using MetaExtractR::join_irf_json().}

\item{shock_size}{A value for the shock size in basis points obtained with MetaExtractR::get_shock_size()}
}
\value{
Annualized shock size.
}
\description{
This function adjusts the shock size based on the periodicity of the interest 
rate. If the periodicity is monthly or quarterly, the function transforms it 
to annualized.
}
