% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_doublechecks.R
\name{assign_doublechecks}
\alias{assign_doublechecks}
\title{Assign doublechecks to a Person}
\usage{
assign_doublechecks(name = NULL, n_doublechecks = 10, weight = 0.001)
}
\arguments{
\item{name}{The researcher ID of the person to whom the doublechecks will be assigned. 
If not specified, the function will stop with an error message. Choose one of "fp", "ph", "me", "dr", "sg".}

\item{n_doublechecks}{The number of doublechecks to assign to the person. Default is 10.}

\item{weight}{The weight the random draw should give to studies for which primary coders did not ask for a doublecheck. Thus, the lower this weight the faster the random draw assigns studies which should be doublechecked according to the primary researcher.}
}
\value{
The function does not return a value but updates a CSV file with new assignments 
and prints a message upon completion.
}
\description{
This function assigns a specified number of studies 
to a person for doublechecks. It updates the list of assigned doublechecks by adding the new 
assignments and saves the updated list in a CSV file.
}
\details{
The function first loads a list of already assigned doublechecks from a CSV file. 
It then retrieves the list of all included studies and their corresponding JSON files. 
The function checks if each included study has a corresponding JSON file and warns if any are missing.
Next, it identifies unassigned studies and assigns the specified number of these studies 
to the provided person's ID. The assignments are then added to the existing list of assignments, 
and the updated list is saved back to the CSV file.
}
\examples{
# Assign 5 IRF folders to a person with the ID "fp"
assign_irfs(name = "fp", n_doublechecks = 10, weight=.001)

}
