% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_irfs.R
\name{assign_irfs}
\alias{assign_irfs}
\title{Assign IRF Folders to a Person}
\usage{
assign_irfs(name = NULL, n_irfs = 5)
}
\arguments{
\item{name}{The name of the person to whom the IRF folders will be assigned. 
If not specified, the function will stop with an error message.}

\item{n_irfs}{The number of IRF folders to assign to the person. Default is 5.}
}
\value{
The function does not return a value but updates a CSV file with new assignments 
and prints a message upon completion.
}
\description{
This function assigns a specified number of Impulse Response Function (IRF) folders 
to a person for data extraction. It updates the list of assigned IRFs by adding the new 
assignments and saves the updated list in a CSV file.
}
\details{
The function first loads a list of already assigned IRFs from a CSV file. 
It then retrieves the list of all current IRF folders and their corresponding JSON files. 
The function checks if each IRF folder has a corresponding JSON file and warns if any are missing.
Next, it identifies unassigned IRF folders and assigns the specified number of these folders 
to the provided person's name. The assignments are then added to the existing list of assignments, 
and the updated list is saved back to the CSV file.
}
\examples{
# Assign 5 IRF folders to a person named "John Doe"
assign_irfs(name = "John Doe", n_irfs = 5)

}
