% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_json_file.R
\name{create_json_file}
\alias{create_json_file}
\title{Create JSON File}
\usage{
create_json_file(
  key,
  rid,
  codebook = "codebook.xlsx",
  folder_path = "data/JSON_files"
)
}
\arguments{
\item{key}{Study key (character): A unique identifier for the study.}

\item{rid}{Researcher ID (character): A unique identifier for the researcher.}

\item{codebook}{File path (character, default = "codebook.xlsx"): The path to the codebook file in Excel format (.xlsx).}

\item{folder_path}{Folder path (character, default = "data/JSON_files"): The folder path where the JSON file will be saved.}
}
\description{
This function generates a JSON file based on a codebook with variable 
names and variable categories, a study key, and a researcher ID.
}
\details{
This function reads the codebook from the specified Excel file, 
checks for uniqueness of variable names, and converts the variable names 
to lowercase. It then creates a JSON file with variable categories as the 
main keys and variable names as subkeys, assigning NA values to each variable.
Additionally, it adds the study key and researcher ID to the generated JSON file, 
as well as placeholder options for categorical variables. Please note that the 
placeholder options are a temporary workaround until they can be listed in 
the codebook itself.
}
\examples{
create_json_file(key = "J4L4L3", rid = "fp")

}
