% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_trans_se_function.R
\name{effect_trans_se_function}
\alias{effect_trans_se_function}
\title{Effect Size Transformation and SE Approximation for Each Respective Case}
\usage{
effect_trans_se_function(d)
}
\arguments{
\item{d}{An observation (i.e. a row) from a dataframe obtained using `join_irf_json()`.}
}
\description{
This function applies the case-dependent effect size transformation and 
standard error approximation based on the provided observation d from a 
dataframe created via the `join_irf_json()` function. The function uses 
many helper functions: `MetaExtractR::join_irf_json()`, 
`MetaExtractR::get_axis_scale()`, `MetaExtractR::get_shock_size()`,
`MetaExtractR::adjust_shock_periodicity()`, `MetaExtractR::get_conf_level()`,
`MetaExtractR::get_crit_val()`.
}
