% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conf_level.R
\name{get_conf_level}
\alias{get_conf_level}
\title{Get Confidence Level from Observation}
\usage{
get_conf_level(d)
}
\arguments{
\item{d}{An observation (a row) from a dataframe obtained using `join_irf_json()`.}
}
\value{
numerical value of the confidence level. For instance, a value of "1_SE" would return 68, and "2_SE" would return 95. 
 Otherwise, the function will attempt to convert the "conf" entry of d to a numeric value.
}
\description{
Extract and convert the confidence level from an observation obtained using `join_irf_json()`.
}
\examples{
# Assuming d is a row of a dataframe from join_irf_json():
# conf_level <- get_conf_level(d)
}
