% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crit_val.R
\name{get_crit_val}
\alias{get_crit_val}
\title{Get Critical value}
\usage{
get_crit_val(confidence_level)
}
\arguments{
\item{confidence_level}{A numeric value between 0 and 1, representing the confidence level.}
}
\value{
A numeric value representing the critical value.
}
\description{
This function computes the critical value given a confidence level.
}
\examples{
get_crit_val(0.68)
get_crit_val(0.95)

}
