% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_irf_json.R
\name{join_irf_json}
\alias{join_irf_json}
\title{Join effect sizes from IRFs and JSON data for a study}
\usage{
join_irf_json(
  key,
  jsonpath = "data/full_text_screening/JSON_files",
  codebook_file = "codebook.csv",
  irf_folder_path = "data/effect_sizes/IRFs/",
  ignore_failed_tests = FALSE
)
}
\arguments{
\item{key}{A study key identifying the study for which IRF data needs to be joined.}

\item{jsonpath}{A character string specifying the path to the folder that contains 
the JSON files.}

\item{codebook_file}{A character string specifiying the path to the codebook.}

\item{irf_folder_path}{A character string specifying the path to the folder 
containing the IRF CSV files.}

\item{ignore_failed_tests}{A logical. If TRUE, the function will parse the 
JSON despite failing consistency tests.}
}
\value{
A data frame containing the joined and processed coding and IRF data.
}
\description{
This function joins effect size data in the form of impulse response functions 
from CSV files with the coding from a JSON file for a given study.
}
\examples{
join_irf_json(
  key = "QSKR5KKX",
  jsonpath = "/Users/franzprante/GitHub/MORPEP/META_CMP/toy_data_extraction_dev/data/JSON_files",
  codebook_file = "/Users/franzprante/GitHub/MORPEP/META_CMP/toy_data_extraction_dev/codebook.csv",
  irf_folder_path = "/Users/franzprante/GitHub/MORPEP/META_CMP/toy_data_extraction_dev/data/effect_sizes/IRFs/",
  ignore_failed_tests = FALSE)
}
