% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_axis_scale_string.R
\name{parse_axis_scale_string}
\alias{parse_axis_scale_string}
\title{Parse Axis Scale String for Scaling Function}
\usage{
parse_axis_scale_string(axis_scale)
}
\arguments{
\item{axis_scale}{A character string specifying the axis scaling, e.g., "y*100", "y/100".
The string should contain a numerical value and an operator (* or /). The function
can also handle some forms of erroneous syntax, such as "y*-1" (paranthesis missing) or "y* / 100" (inserted spaces).}
}
\value{
A list containing: axis_scaling_function: A function to apply the desired scaling on y-values.; 
scaling_parameter: A numeric value indicating the scaling factor or multiplier.
}
\description{
This function parses a given axis scaling string (created within the get_axis_scale 
function) to extract the numerical value and operator for scaling purposes. 
It returns a list containing a scaling function (axis_scale_fun) and the associated scaling parameter (y_scale).
}
\examples{
# General scaling for all variables
parse_axis_scale_string("y*100")
parse_axis_scale_string("y/100")
parse_axis_scale_string("y*(-1)")
parse_axis_scale_string("y*-1") 
# Even some erroneous coding syntax (no parenthesis around "-1") is correctly parsed, but "y*- 1" would not work, for example.
parse_axis_scale_string("y*0.01")
parse_axis_scale_string("y / 100") # Even some erroneous coding syntax (spaces around "/") is correctly parsed
# Specific scaling for one variable
parse_axis_scale_string("log_rgdp_y*100")
parse_axis_scale_string("log_cpi_y/100")
parse_axis_scale_string("log_rgdp_y*(-1)")
parse_axis_scale_string("logdiff_ip_y*0.1")
parse_axis_scale_string("log_rgdpy*100") # Even some erroneous coding syntax ( no "_" left of "y") is correctly parsed

}
