% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_study_json.R
\name{parse_study_json}
\alias{parse_study_json}
\title{Parse a JSON File Containing Study Data into a Dataframe}
\usage{
parse_study_json(
  json_file,
  json_path,
  codebook = "codebook.csv",
  view_data = TRUE
)
}
\arguments{
\item{json_file}{The name of the JSON file to be parsed.}

\item{json_path}{The path to the JSON file to be parsed.}

\item{codebook}{The path to the codebook csv file, with default value 'codebook.csv'.}

\item{view_data}{Logical indicating whether to view the resulting dataframe, with default TRUE.}
}
\value{
The function returns a dataframe with study data. Variables in the dataframe 
   correspond to variables in the JSON file and align with the codebook. Each row 
   in the dataframe corresponds to a model from the JSON file. Values from the JSON 
   file that are "null" are converted to NA in the dataframe.
}
\description{
This function is designed to import a JSON file that contains study data,
cross-check this data with a specified codebook, and return the parsed data
as a dataframe. The function will stop and throw an error message if it 
encounters inconsistencies between the JSON data and the codebook, or 
inconsistencies in the number of models per variable in the JSON data.
}
\examples{
parse_study_json(
 json_file = "dev_filled.json", 
 json_path = "/Users/franzprante/GitHub/MORPEP/META CMP/toy_data_extraction_dev/data/JSON_files", 
 codebook = "/Users/franzprante/GitHub/MORPEP/META CMP/data/codebook.csv", 
 view_data = FALSE)

}
