% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_json_file.R
\name{create_json_file}
\alias{create_json_file}
\title{Create JSON File}
\usage{
create_json_file(
  key,
  rid,
  codebook = "codebook.csv",
  folder_path = "data/JSON_files"
)
}
\arguments{
\item{key}{Study key (character): A unique identifier for the study.}

\item{rid}{Researcher ID (character): A unique identifier for the researcher.}

\item{codebook}{File path (character, default = "codebook.csv"): The path to the codebook file in csv format.}

\item{folder_path}{Folder path (character, default = "data/JSON_files"): The folder path where the JSON file will be saved.}
}
\description{
This function generates a JSON file based on a codebook with variable 
names and variable categories, a study key, and a researcher ID.
}
\details{
This function reads the codebook from the specified csv file, 
converts the variable names to lowercase and checks for uniqueness of 
variable names. It then creates a JSON file with variable categories as the 
main keys and variable names as subkeys, assigning NA values to each variable.
For variables that are indicated as "factor" in the codebook (in column "type"),
the function automatically sets the variable's value to the options in the 
codebook (in column "entries"). Additionally, it adds the study key and 
researcher ID to the generated JSON file.
}
\examples{
create_json_file(
 key = "dev", 
 rid = "fp", 
 folder_path = "/Users/franzprante/GitHub/MORPEP/META_CMP/toy_data_extraction_dev/data/JSON_files", 
 codebook = "/Users/franzprante/GitHub/MORPEP/META_CMP/data/codebook.csv")

}
